# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "description": "Siges/swagger.json",
        "version": "1.0"
    },
    "paths": {
...
        "/api/coord": {
            "get": {
                "summary": "Retornar uma lista de coordenadores.",
                "operationId": "324045f50c4920cb1810b1ecd94b37b6",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página para a paginação",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retornar uma lista de coordenadores."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/coord/{user_id}": {
            "get": {
                "summary": "Retorna detalhes do coordenador por ID.",
                "description": "Este endpoint permite buscar coordenadores associados a um determinado ID de usuário.",
                "operationId": "7e6ee9289134f49231d151cee2e0a34c",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID do usuário para buscar coordenadores.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes dos coordenadores encontrados."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/coord/{id}": {
            "put": {
                "summary": "Associa um usuário a uma coordenadoria (CDE) com base no ID do usuario.",
                "operationId": "93ad6c2cc974da397af3d80e6a060b8b",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID do usuário a ser associado à coordenadoria.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "coordenadoria": {
                                        "type": "string",
                                        "example": "Coordenadoria Exemplo"
                                    },
                                    "municipio": {
                                        "type": "string",
                                        "example": "O Município é obrigatório apenas se a coordenadoria for regional"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário associado a uma CDE com sucesso."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/import-coord": {
            "post": {
                "summary": "Importa novos coordenadores a partir de um arquivo CSV.",
                "operationId": "ddc9932ecbb350244cb13d535d29532c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "Arquivo CSV contendo os dados dos coordenadores.",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registros importados com sucesso."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/export-coord": {
            "get": {
                "summary": "Export csv coordenadores.",
                "operationId": "4c6d5896ced79afe1c588cae4f57e75b",
                "responses": {
                    "200": {
                        "description": "Export csv coordenadores.",
                        "content": {
                            "application/csv": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Adicione o token JWT no cabeçalho da requisição. Exemplo: 'Authorization: Bearer {token}'",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}